+++ app/league/api/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+"""The API module."""
+from . import views  # noqa
