+++ app/league/api/forms.py
@@ -8,35 +8,6 @@ from wtforms.validators import DataRequired, NumberRange
 from league.models import Color, Game, Player
 
 
-class PlayerCreateForm(FlaskForm):
-    """Player creation form."""
-
-    first_name = StringField('first_name', validators=[DataRequired()])
-    last_name = StringField('last_name', validators=[DataRequired()])
-    aga_id = IntegerField(
-        'aga_id', validators=[NumberRange(0, 50000)])
-    aga_rank = IntegerField(
-        'aga_rank', validators=[NumberRange(-30, 9)])
-
-    @staticmethod
-    def validate_aga_rank(form, field):
-        """Check that the input rank is an acceptable aga rank."""
-        if field.data < -30 or field.data > 9 or field.data == 0:
-            raise ValidationError('Rank must be -30 to -1 or 1 to 9')
-
-
-class PlayerDeleteForm(FlaskForm):
-    """Player deletion form."""
-
-    player_id = IntegerField('player_id', validators=[NumberRange(0, 50000)])
-
-    @staticmethod
-    def validate_player_id(form, field):
-        """Check that players are not in extant games."""
-        if len(Player.get_by_id(field.data).games) > 0:
-            raise ValidationError('Players with extant games cannot be deleted')
-
-
 class GameCreateForm(FlaskForm):
     """Game creation form."""
 
@@ -100,10 +71,3 @@ class GameUpdateForm(FlaskForm):
         """Check that IDs are different."""
         if form.black_id.data == form.white_id.data:
             raise ValidationError('Players cannot play themselves')
-
-
-class ReportGenerateForm(FlaskForm):
-    """Report generation form."""
-
-    season = IntegerField('season', validators=[NumberRange(1, 10000)])
-    episode = IntegerField('episode', validators=[NumberRange(1, 10000)])
