+++ app/tests/__init__.py
@@ -1,1 +1,2 @@
 """Tests for the app."""
+from . import factories  # noqa
