+++ app/tests/conftest.py
@@ -9,7 +9,7 @@ from league.app import create_app
 from league.database import db as _db
 from league.settings import TestConfig
 
-from .factories import GameFactory, PlayerFactory, UserFactory
+from .factories import UserFactory
 
 
 @pytest.yield_fixture(scope='function')
@@ -71,31 +71,3 @@ def users(db):
     users = [UserFactory(), UserFactory()]
     db.session.commit()
     return users
-
-
-@pytest.fixture
-def players(db):
-    """Some players for the tests."""
-    players = [PlayerFactory(), PlayerFactory()]
-    db.session.commit()
-    return players
-
-
-@pytest.fixture
-def games(db):
-    """Some games for the tests."""
-    games = [GameFactory(), GameFactory()]
-    db.session.commit()
-    return games
-
-
-@pytest.fixture
-def season_choices():
-    """Season choices to use when testing game create form."""
-    return [(s, s) for s in range(0, 3)]
-
-
-@pytest.fixture
-def episode_choices():
-    """Episode choices to use when testing game create form."""
-    return [(e, e) for e in range(0, 2)]
