+++ app/tests/dashboard/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for Dashboard pages."""
