+++ app/tests/factories.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
 """Factories to help in tests."""
-from factory import PostGenerationMethodCall, Sequence, SubFactory
+from factory import PostGenerationMethodCall, Sequence
 from factory.alchemy import SQLAlchemyModelFactory
 
 from league.admin.models import User
-from league.dashboard.models import Color, Game, Player
 from league.database import db
 
 
@@ -32,35 +31,3 @@ class UserFactory(BaseFactory):
         """Factory configuration."""
 
         model = User
-
-
-class PlayerFactory(BaseFactory):
-    """Player factory."""
-
-    first_name = Sequence(lambda n: 'first{0}'.format(n))
-    last_name = Sequence(lambda n: 'last{0}'.format(n))
-    aga_id = Sequence(lambda n: n + 1000)
-    aga_rank = Sequence(lambda n: (n % 9) + 1)
-
-    class Meta:
-        """Factory configuration."""
-
-        model = Player
-
-
-class GameFactory(BaseFactory):
-    """Game factory."""
-
-    white = SubFactory(PlayerFactory)
-    black = SubFactory(PlayerFactory)
-    winner = Color.white
-    handicap = 0
-    komi = 7
-    season = 1
-    episode = 1
-    created_at = None
-
-    class Meta:
-        """Factory configuration."""
-
-        model = Game
