+++ app/tests/public/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for public pages."""
